requirements.txt
flask==3.1.1
transformers==4.52.4
torch==2.7.1
pillow==11.2.1
numpy==2.3
pytesseract==0.3.13
python-dotenv==1.1.0
accelerate==1.7.0
bitsandbytes==0.46.0
sentencepiece==0.2.0
protobuf==6.31.1
einops==0.8.1
chromadb==1.0.12
langchain==0.3.25
pypdf==5.6.0
sentence-transformers==4.1.0
langchain-community==0.3.25
opencv-python==4.11.0.86
flash-attn==2.5.6
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
werkzeug==3.0.1